requirements.txt
Flask==2.3.3
tensorflow==2.15.1
numpy==1.21.0
Pillow==8.3.1
Werkzeug==2.3.7
bcrypt==4.0.1
pdfkit==1.0.0 